+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Please follow the format in [Keep a Changelog](http://keepachangelog.com/)
 
 - Support for ActiveRecord 4.2.x
 - Support for Mysql2 4.x
+- Allow passing any `pt-online-schema-change`'s arguments through the
+   `PERCONA_ARGS` env var when executing a migration with `rake db:migrate:up`
+   or `db:migrate:down`.
 
 ## [1.0.0] - 2016-11-30